move config logic into version specific generators
